update typescript and refactor code to resolve compile errors
